updates documentation and project structure for src layout with console scripts
