+++ src/dream_cli/__main__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+"""Entry point for `python -m dream_cli` (interactive CLI)."""
+from dream_cli._scripts.dreamcli import main
+
+main()
